% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plotCommNetwork}
\alias{plotCommNetwork}
\alias{plotCommNetwork,SOTK-method}
\title{Generate community network plot}
\usage{
\S4method{plotCommNetwork}{SOTK}(
  object,
  vertexInfo,
  filename = "community.pdf",
  width = 10,
  height = 10
)
}
\arguments{
\item{object}{A \code{SOTK} object}

\item{vertexInfo}{\code{list} Pie annotation that includes node label, size, pie (proportion), and legend.}

\item{filename}{\code{character} Output file name}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Correlation network plot
}
\description{
Generate community network plot
}
\examples{
\dontrun{
plotCommNetwork(
    object     = soObj,
    vertexInfo = list(
        vertexLabel = vertexLabel,
        vertexSize  = vertexSize,
        vertexPie   = vertexPie,
        legend      = legend,
        legendCol   = legendCol
    )
)
}

}
