% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plotCorrDensity}
\alias{plotCorrDensity}
\alias{plotCorrDensity,SOSet-method}
\title{Generate correlation coefficient density plot}
\usage{
\S4method{plotCorrDensity}{SOSet}(
  object,
  filename = "pairwiseCorrDensityPlot.pdf",
  breaks = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{object}{A \code{SOSet} object}

\item{filename}{\code{character} Output file name. If \code{NULL}, the plot
is drawn on the active graphics device.}

\item{breaks}{\code{numeric} vector of histogram break points passed to
\code{hist()}. Defaults to \code{seq(-1, 1, 0.1)} when \code{NULL}.}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Coefficient density plot
}
\description{
Generate correlation coefficient density plot
}
\examples{
\dontrun{
plotCorrDensity(soSet)
}

}
