% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,SOTK-method}
\title{Generate correlation network plot}
\usage{
\S4method{plotNetwork}{SOTK}(
  object,
  weighted = TRUE,
  label = FALSE,
  annot = "cohort",
  vertexSize = 5,
  vertexLabelCex = 1,
  edgeAlpha = 0.2,
  filename = "network.pdf",
  width = 10,
  height = 10
)
}
\arguments{
\item{object}{A \code{SOTK} object}

\item{weighted}{\code{logical} Whether generate a correlation network with weighted or unweighted layout. Default is TRUE.}

\item{label}{\code{logical} Whether annotate metagene label or not. Default is FALSE.}

\item{annot}{\code{character} Whether data or community annotation on the network. Default is data.}

\item{vertexSize}{\code{numeric} Vertex/node size on the network. Default is 5.}

\item{vertexLabelCex}{\code{numeric} Label font size on the nodes. Default is 1.}

\item{edgeAlpha}{\code{numeric} Alpha value for the edge transparency. Default is 0.2.}

\item{filename}{\code{character} Output file name}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Correlation network plot
}
\description{
Generate correlation network plot
}
\examples{
\dontrun{
plotNetwork(object = soObj)
}

}
