% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{selectRank}
\alias{selectRank}
\alias{selectRank,SOTK-method}
\title{Generate community network stats}
\usage{
\S4method{selectRank}{SOTK}(
  object,
  vertexSize = 8,
  vertexLabelCex = 0.75,
  edgeAlpha = 0.2,
  filename = "Coverage.pdf",
  width = 10,
  height = 10
)
}
\arguments{
\item{object}{A \code{SOTK} object}

\item{vertexSize}{\code{numeric} Vertex/node size on the network. Default is 8.}

\item{vertexLabelCex}{\code{numeric} Label font size on the nodes. Default is 0.75.}

\item{edgeAlpha}{\code{numeric} Alpha value for the edge transparency. Default is 0.2.}

\item{filename}{\code{character} Output file name}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Metagene marking at each rank on the community-weighted layout
}
\description{
Generate community network stats
}
\examples{
\dontrun{
selectRank(object = soObj)
}

}
